feat: 🎸 add archive folder and test hook props
