+++ client-react/src/components/archive/PluginCameraItem.js
