+++ client-react/src/components/archive/PluginSwitchButton.js
